Tests: ensure models contain only one record per file
