+++ tests/Relation/OneToOneTestCase.php
@@ -30,14 +30,14 @@
  * @since       1.0
  * @version     $Revision$
  */
-class Doctrine_Relation_OneToOne_TestCase extends Doctrine_UnitTestCase 
+class Doctrine_Relation_OneToOne_TestCase extends Doctrine_UnitTestCase
 {
-    public function prepareData() 
+    public function prepareData()
     { }
-    public function prepareTables() 
-    { 
-        $this->tables = array('gnatUser','gnatEmail','Email','Entity','Record_City', 'Record_Country', 'SelfRefTest');
-        
+    public function prepareTables()
+    {
+        $this->tables = array('GnatUser','GnatEmail','Email','Entity','Record_City', 'Record_Country', 'SelfRefTest');
+
         parent::prepareTables();
     }
 
@@ -46,27 +46,27 @@ class Doctrine_Relation_OneToOne_TestCase extends Doctrine_UnitTestCase
         $city = new Record_City();
         $country = $city->Country;
 
-        $this->assertTrue($country instanceof Record_Country);  
+        $this->assertTrue($country instanceof Record_Country);
     }
-    
+
     public function testSelfReferentialOneToOneRelationsAreSupported()
     {
         $ref = new SelfRefTest();
-        
+
         $rel = $ref->getTable()->getRelation('createdBy');
 
         $this->assertEqual($rel->getForeign(), 'id');
         $this->assertEqual($rel->getLocal(), 'created_by');
-        
+
         $ref->name = 'ref 1';
         $ref->createdBy->name = 'ref 2';
-        
+
         $ref->save();
     }
     public function testSelfReferentialOneToOneRelationsAreSupported2()
     {
         $this->connection->clear();
-        
+
         $ref = $this->conn->queryOne("FROM SelfRefTest s WHERE s.name = 'ref 1'");
         $this->assertEqual($ref->name, 'ref 1');
         $this->assertEqual($ref->createdBy->name, 'ref 2');
@@ -88,14 +88,14 @@ class Doctrine_Relation_OneToOne_TestCase extends Doctrine_UnitTestCase
 
     public function testSavingRelatedObjects()
     {
-        $user = new gnatUser();
+        $user = new GnatUser();
         $user->name = 'test';
-        $email = new gnatEmail();
+        $email = new GnatEmail();
         $email->address = 'test3@test.com';
         $user->Email = $email;
         $user->save();
-        $this->assertTrue($user->Email instanceOf gnatEmail);
+        $this->assertTrue($user->Email instanceOf GnatEmail);
         $this->assertEqual($user->foreign_id, $user->Email->id);
-        
+
     }
 }
