+++ tests/models/Blog.php
@@ -10,23 +10,3 @@ class Blog extends Doctrine_Record
         $this->actAs('Taggable');
     }
 }
-class Taggable extends Doctrine_Template
-{
-    public function setUp()
-    {
-        //$this->hasMany('[Component]TagTemplate as Tag');
-    }
-}
-class TagTemplate extends Doctrine_Record
-{
-    public function setTableDefinition()
-    {
-        $this->hasColumn('name', 'string', 100);
-        $this->hasColumn('description', 'string');
-    }
-
-    public function setUp()
-    {
-        //$this->hasOne('[Component]', array('onDelete' => 'CASCADE'));
-    }
-}
