+++ tests/models/GnatEmail.php
@@ -1,10 +1,8 @@
 <?php
-class gnatEmail extends Doctrine_Record 
+class GnatEmail extends Doctrine_Record
 {
     public function setTableDefinition() 
     {
         $this->hasColumn('address', 'string', 150);
     }
-    
-    
 }
