+++ tests/models/GnatUserTable.php
@@ -1,0 +1,3 @@
+<?php 
+
+class GnatUserTable { }
