+++ tests/models/GroupTable.php
@@ -1,0 +1,5 @@
+<?php
+
+// grouptable doesn't extend Doctrine_Table -> Doctrine_Connection
+// won't initialize grouptable when Doctrine_Connection->getTable('Group') is called
+class GroupTable { }
