+++ tests/models/MigrationPhonenumber.php
@@ -1,0 +1,10 @@
+<?php
+
+class MigrationPhonenumber extends Doctrine_Record
+{
+    public function setTableDefinition()
+    {
+        $this->hasColumn('user_id', 'integer');
+        $this->hasColumn('phonenumber', 'string', 255);
+    }
+}
