+++ tests/models/MigrationProfile.php
@@ -1,0 +1,9 @@
+<?php
+
+class MigrationProfile extends Doctrine_Record
+{
+    public function setTableDefinition()
+    {
+        $this->hasColumn('name', 'string', 255);
+    }
+}
