+++ tests/models/MigrationUser.php
@@ -1,0 +1,10 @@
+<?php
+
+class MigrationUser extends Doctrine_Record
+{
+    public function setTableDefinition()
+    {
+        $this->hasColumn('username', 'string', 255);
+        $this->hasColumn('password', 'string', 255);
+    }
+}
