+++ tests/models/Taggable.php
@@ -1,0 +1,8 @@
+<?php
+class Taggable extends Doctrine_Template
+{
+    public function setUp()
+    {
+        //$this->hasMany('[Component]TagTemplate as Tag');
+    }
+}
